# Reduce the number of commits in db:seed for better performance