Before

```
% bin/rails db:seed
% psql -c 'delete from episodes' annict_development
DELETE 1565
% time bin/rails db:seed
bin/rails db:seed  26.08s user 1.50s system 36% cpu 1:16.16 total
```

After

```
% bin/rails db:seed
% psql -c 'delete from episodes' annict_development
DELETE 1565
% time bin/rails db:seed
bin/rails db:seed  13.77s user 0.90s system 87% cpu 16.805 total
````